File walk function working well

+++ packages/rust/conundrum/src/lang/lib/general/pagination/pagination_params.rs
@@ -4,6 +4,6 @@ use typeshare::typeshare;
 #[typeshare]
 #[derive(Debug, Serialize, Deserialize, uniffi::Record)]
 pub struct PaginationParams {
-    pub per_page: u32,
-    pub page: u32,
+    pub per_page: u64,
+    pub page: u64,
 }
